# docs/swagger.yaml
basePath: /api
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_handlers.AnniversaryResponse:
    properties:
      createdAt:
        type: string
      date:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      group_id:
        type: integer
      id:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  internal_handlers.CreateAnniversaryRequest:
    properties:
      date:
        type: string
      title:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - date
    - title
    type: object
...
  internal_handlers.CreateNoteRequest:
    properties:
      color:
        maxLength: 20
        type: string
      content:
        maxLength: 1000
        type: string
      show_at:
        type: string
      type:
        type: string
    required:
    - content
    - type
    type: object
  internal_handlers.CreatePhotoRequest:
    properties:
      description:
        maxLength: 500
        type: string
      url:
        maxLength: 1024
        type: string
    required:
    - url
    type: object
  internal_handlers.CreateWishlistRequest:
    properties:
      content:
        maxLength: 255
        type: string
    required:
    - content
    type: object
...
  internal_handlers.NoteResponse:
    properties:
      author:
        $ref: '#/definitions/vocalin-backend_internal_models.User'
      author_id:
        type: integer
      color:
        type: string
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      group_id:
        type: integer
      id:
        type: integer
      is_burned:
        type: boolean
      show_at:
        type: string
      type:
        description: '"normal", "burn", "timed"'
        type: string
      updatedAt:
        type: string
    type: object
  internal_handlers.PhotoResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/vocalin-backend_internal_models.Comment'
        type: array
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      likes:
        items:
          $ref: '#/definitions/vocalin-backend_internal_models.Like'
        type: array
      updatedAt:
        type: string
      uploader_id:
        type: integer
      url:
        type: string
    type: object
...
  internal_handlers.WishlistResponse:
    properties:
      completed_at:
        type: string
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      group_id:
        type: integer
      id:
        type: integer
      is_completed:
        type: boolean
      updatedAt:
        type: string
    type: object
  vocalin-backend_internal_models.Comment:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      photo_id:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/vocalin-backend_internal_models.User'
      user_id:
        type: integer
    type: object
...
  vocalin-backend_internal_models.Like:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      photo_id:
        type: integer
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  vocalin-backend_internal_models.User:
    properties:
      avatar_url:
        type: string
      createdAt:
        type: string
      current_status:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      group_id:
        description: Nullable if not in a group
        type: integer
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      status_updated_at:
        type: string
      updatedAt:
        type: string
      wechat_id:
        type: string
    type: object
  vocalin-backend_internal_response.APIResponse:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      meta: {}
    type: object
  vocalin-backend_internal_response.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Vocalin（窝聚）后端服务，基于 Gin + GORM + Viper + Zap + Validator + JWT + Swagger
    构建
  title: Vocalin API
  version: "2.0"
paths:
...
  /profile/anniversaries:
    get:
      parameters:
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handlers.AnniversaryResponse'
                  type: array
                meta:
                  $ref: '#/definitions/vocalin-backend_internal_response.PaginationMeta'
              type: object
      security:
      - BearerAuth: []
      summary: 获取纪念日列表
      tags:
      - Profile
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Anniversary Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateAnniversaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.AnniversaryResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 新增纪念日
      tags:
      - Profile
  /profile/export:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 导出个人数据
      tags:
      - Profile
  /profile/leave:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 退出当前空间
      tags:
      - Profile
  /records/notes:
    get:
      parameters:
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handlers.NoteResponse'
                  type: array
                meta:
                  $ref: '#/definitions/vocalin-backend_internal_response.PaginationMeta'
              type: object
      security:
      - BearerAuth: []
      summary: 获取便签列表
      tags:
      - Records
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Note Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.NoteResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建便签
      tags:
      - Records
  /records/photos:
    get:
      parameters:
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handlers.PhotoResponse'
                  type: array
                meta:
                  $ref: '#/definitions/vocalin-backend_internal_response.PaginationMeta'
              type: object
      security:
      - BearerAuth: []
      summary: 获取照片列表
      tags:
      - Records
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Photo Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreatePhotoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.PhotoResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 上传照片记录
      tags:
      - Records
  /records/wishlist:
    get:
      parameters:
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handlers.WishlistResponse'
                  type: array
                meta:
                  $ref: '#/definitions/vocalin-backend_internal_response.PaginationMeta'
              type: object
      security:
      - BearerAuth: []
      summary: 获取愿望清单
      tags:
      - Records
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Wishlist Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateWishlistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vocalin-backend_internal_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handlers.WishlistResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 新增愿望清单项
      tags:
      - Records
  /records/wishlist/{id}/complete:
    put:
      parameters:
      - description: Wishlist Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.WishlistResponse'
      security:
      - BearerAuth: []
      summary: 完成愿望清单项
      tags:
      - Records
  /records/wishlist/{id}/incomplete:
    put:
      parameters:
      - description: Wishlist Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.WishlistResponse'
      security:
      - BearerAuth: []
      summary: 取消完成愿望清单项
      tags:
      - Records
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer <token> 传递 JWT 访问令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"